Add llb.Platform when building llb.Image
- Refactor solve tree to output branches under mount and add test for multi platform llb
